Some tidying of co-ordinate handling.

+++ src/canvasutil.rs
@@ -56,14 +56,14 @@ pub fn aspect_ratio(canvas: &CanvasElement) -> f32 {
 #[derive(Clone,Eq,PartialEq,Hash)]
 pub struct FCFont {
     spec: String,
-    height: u32,
-    xpad: u32,
-    ypadtop: u32,
-    ypadbot: u32
+    height: i32,
+    xpad: i32,
+    ypadtop: i32,
+    ypadbot: i32
 }
 
 impl FCFont {
-    pub fn new(size : u32,family: &str) -> FCFont {
+    pub fn new(size : i32,family: &str) -> FCFont {
         FCFont { spec: format!("{}px {}",size,family),
                  height: size, ypadtop: 0, ypadbot: 4, xpad: 0 }
     }
@@ -76,17 +76,17 @@ impl FCFont {
 pub struct FlatCanvas {
     canvas: CanvasElement,
     context : CanvasRenderingContext2d,
-    width: u32,
-    height: u32,
+    width: i32,
+    height: i32,
 }
 
 impl FlatCanvas {
-    pub fn create(width: u32,height: u32) -> FlatCanvas {
+    pub fn create(width: i32,height: i32) -> FlatCanvas {
         let canvas_holder = domutil::query_select("#managedcanvasholder");
         let canvas : CanvasElement = document().create_element("canvas").ok().unwrap().try_into().unwrap();
         canvas_holder.append_child(&canvas);
-        canvas.set_width(width);
-        canvas.set_height(height);
+        canvas.set_width(width as u32);
+        canvas.set_height(height as u32);
         let context : CanvasRenderingContext2d = canvas.get_context().unwrap();
         context.set_fill_style_color("white");
         context.fill_rect(0.,0.,width as f64,height as f64);
@@ -94,19 +94,19 @@ impl FlatCanvas {
         FlatCanvas { canvas, context, height, width }
     }
     
-    pub fn text(&self,text : &str,x : u32, y: u32, font: &FCFont, col: &Colour) -> (u32,u32) {
+    pub fn text(&self,text : &str,x : i32, y: i32, font: &FCFont, col: &Colour) -> (i32,i32) {
         font.setup(&self.context);
         self.context.set_text_baseline(TextBaseline::Top);
         self.context.set_fill_style_color(&col.to_css()[..]);
         self.context.set_stroke_style_color(&col.to_css()[..]);
         self.context.fill_text(text,(x+font.xpad).into(),(y+font.ypadtop).into(),None);
         let m = self.context.measure_text(text);
-        let width_px = m.unwrap().get_width().ceil() as u32;
+        let width_px = m.unwrap().get_width().ceil() as i32;
         let height_px = font.height;
         (width_px+2*font.xpad,height_px+font.ypadtop+font.ypadbot)
     }
     
-    pub fn bitmap(&self, data: &Vec<u8>, x: u32, y: u32, width: u32, height: u32) {
+    pub fn bitmap(&self, data: &Vec<u8>, x: i32, y: i32, width: i32, height: i32) {
         let pixels: TypedArray<u8> = data[..].into();
         js! {
             var id = @{&self.context}.createImageData(@{width},@{height});
@@ -115,10 +115,10 @@ impl FlatCanvas {
         };
     }
 
-    pub fn measure(&self,text : &str, font: &FCFont) -> (u32,u32) {
+    pub fn measure(&self,text : &str, font: &FCFont) -> (i32,i32) {
         font.setup(&self.context);
         let m = self.context.measure_text(text);
-        let width_px = m.unwrap().get_width().ceil() as u32;
+        let width_px = m.unwrap().get_width().ceil() as i32;
         let height_px = font.height;
         (width_px+2*font.xpad,height_px+font.ypadtop+font.ypadbot)
     }
@@ -127,11 +127,11 @@ impl FlatCanvas {
         &self.canvas
     }
     
-    pub fn prop_x(&self,x: u32) -> f32 {
+    pub fn prop_x(&self,x: i32) -> f32 {
         (x as f64 / self.width as f64) as f32
     }
 
-    pub fn prop_y(&self,y: u32) -> f32 {
+    pub fn prop_y(&self,y: i32) -> f32 {
         (y as f64 / self.height as f64) as f32
     }
